Script gathers all files for IGV.

+++ prepIGV.sh
@@ -1,34 +1,42 @@
 #!/bin/bash
+set -eu -o pipefail
 PROGRAM=$(basename $0)
 gethelp=false
 scaffolds=false
 transcripts=false
 gff=false
-while getopts :hs:t:g: opt
-do
-	case $opt in
-		h) gethelp=true;;
-		s) scaffolds=$OPTARG; if [ ! -s "$scaffolds" ]; then echo "$(basename $scaffolds) is an invalid file."1>&2;fi;exit 1;;
-		t) transcripts=$OPTARG; if [ ! -s "$transcripts" ]; then echo "$(basename $transcripts) is an invalid file."1>&2;fi;exit 1;;
-		g) gff=$OPTARG; if [ ! -s "$gff" ]; then echo "$(basename $gff) is an invalid file." 1>&2;fi;exit 1;;
-		\?) echo "$PROGRAM: Invalid option $OPTARG" 1>&2; exit 1;;
-	esac
-done
-shift $((OPTIND-1))
-
-# Write help/error messages here
-if [[ "$#" -ne 0 || "$gethelp" = true ]]
+if [[ "$#" -eq 0 ]]
 then
 	echo "USAGE: $PROGRAM -s <SCAFFOLD FASTA> -t <TRANSCRIPT FASTA> -g <GFF file>" 1>&2 
 	echo "DESCRIPTION: Fetches necessary files for IGV." 1>&2
 	exit 1
 fi
 
+
+while getopts :hs:t:g: opt
+do
+	case $opt in
+		h) gethelp=true;;
+		s) scaffolds=$OPTARG; if [ ! -s "$scaffolds" ]; then echo "$(basename $scaffolds) is an invalid file."1>&2;exit 1;fi;;
+		t) transcripts=$OPTARG; if [ ! -s "$transcripts" ]; then echo "$(basename $transcripts) is an invalid file."1>&2;exit 1; fi;;
+		g) gff=$OPTARG; if [ ! -s "$gff" ]; then echo "$(basename $gff) is an invalid file." 1>&2;exit 1; fi;;
+		\?) echo "$PROGRAM: Invalid option $OPTARG" 1>&2; exit 1;;
+	esac
+done
+shift $((OPTIND-1))
+# Write help/error messages here
+if [[ "$gethelp" = true ]]
+then
+	echo "USAGE: $PROGRAM -s <SCAFFOLD FASTA> -t <TRANSCRIPT FASTA> -g <GFF file>" 1>&2 
+	echo "DESCRIPTION: Fetches necessary files for IGV." 1>&2
+	exit 1
+fi
 # if scaffolds, gffs and transcripts are given-- typical case or proteome jackhmmer spruce
 if [[ "$scaffolds" != false && "$transcripts" != false && "$gff" != false  ]]
 then
+	echo "Scaffolds, transcripts and GFF files detected!" 1>&2
 	# GET scaffolds, gffs and transcripts
-	echo "Fetching scaffolds, transcripts and GFF files..."
+	echo "Fetching scaffolds, transcripts and GFF files..." 1>&2
 	if [ ! -e "scaffolds" ]
 	then
 		mkdir scaffolds
@@ -66,7 +74,7 @@ then
 			cd ..
 		fi
 		seqtk subseq $transcripts <(echo $i) >> transcripts/${scaffold}.transcripts.fa
-		grep $i $gff >> gffs/${scaffold}.gff
+		grep ${i::-7} $gff >> gffs/${scaffold}.gff
 	done
 
 	date=$(date | awk '{if($3<10) {print "0" $3 $2 $6} else {print $3 $2 $6}}')
@@ -77,14 +85,15 @@ then
 	cd ../gffs
 	cat *.gff > ../all.${date}.gff
 	cd ..
+	echo "...Done." 1>&2
 
-
-# if for spruce transcriptome, where transcripts and tbl are given
+# if for spruce transcriptome, where transcripts
 elif [[ "$transcripts" != false && "$scaffolds" = false && "$gff" = false ]]
 then
+	echo "Transcripts detected!" 1>&2
 	# in this case, must MAKE a gff.
 	echo "Fetching transcripts..." 1>&2
-	echo "Making GFF..." 1>&2 
+	echo "Making GFFs..." 1>&2 
 	mkdir -p transcripts
 	mkdir -p gffs
 	mkdir -p igv
@@ -94,7 +103,10 @@ then
 		transcript_name=$(echo $i | awk '{print $1}' | awk -F "_" '{print $2}' | awk -F ":" '{print $1}')
 
 		# Get transcript sequence and deposit into the folder
-		seqtk subseq $transcripts <(echo $transcript_name) > transcripts/${transcript_name}.transcript.fa
+		if [[ ! -e "transcripts/${transcript_name}.transcript.fa" ]]
+		then
+			seqtk subseq $transcripts <(echo $transcript_name) > transcripts/${transcript_name}.transcript.fa
+		fi
 		length=$(seqtk comp transcripts/${transcript_name}.transcript.fa | awk '{print $2}')
 
 
@@ -109,11 +121,14 @@ then
 		
 		# Start making GFF
 		# First line of GFF
-		echo -e "##gff-version 3\n##sequence-region $transcript_name 1 $length" > gffs/${transcript_name}.gff
-		echo -e "$transcript_name\tORFfinder\tgene\t1\t$length\t.\t$strand\t.\tID=${transcript_name}" >> gffs/${transcript_name}.gff
+		if [[ ! -e "gffs/${transcript_name}.gff" ]]
+		then
+			echo -e "##gff-version 3\n##sequence-region $transcript_name 1 $length" > gffs/${transcript_name}.gff
+		fi
+	#	echo -e "$transcript_name\tORFfinder\tgene\t1\t$length\t.\t$strand\t.\tID=${transcript_name}" >> gffs/${transcript_name}.gff
 		echo -e "$transcript_name\tORFfinder\tmRNA\t1\t$length\t.\t$strand\t.\tID=${transcript_name}-mRNA-1;Parent=${transcript_name}" >> gffs/${transcript_name}.gff
 		echo -e "$transcript_name\tORFfinder\texon\t$cds_begin\t$cds_end\t.\t$strand\t0\tID=${transcript_name}:exon;Parent=${transcript_name}-mRNA-1" >> gffs/${transcript_name}.gff
-		echo -e "$transcript_name\tORFfinder\tCDS\t$cds_begin\t$cds_end\t.\t$strand\t0\tID=${transcript_name}:cds;Parent=${transcript_name}-mRNA-1" >> gffs/${transcrip_name}.gff 
+		echo -e "$transcript_name\tORFfinder\tCDS\t$cds_begin\t$cds_end\t.\t$strand\t0\tID=${transcript_name}:cds;Parent=${transcript_name}-mRNA-1" >> gffs/${transcript_name}.gff 
 		cd igv
 		ln -sf ../transcripts/${transcript_name}.transcript.fa
 		ln -sf ../gffs/${transcript_name}.gff
@@ -125,8 +140,10 @@ then
 	cd ../gffs
 	cat *.gff > ../all.${date}.gff
 	cd ..
+	echo "...Done." 1>&2
 elif [[ "$transcripts" != false && "$scaffolds" = false && "$gff" != false ]]
 then
+	echo "Transcripts and GFF detected!" 1>&2
 	# necessary GFF headers are in line 1 and 7 of NCBI GFFs
 	awk 'NR==1 || NR==7' $gff > amp.gff
 	# NCBI data is named differently
@@ -138,6 +155,7 @@ then
 		# Fetch GFF
 		grep $i $gff >> amp.gff
 	done
+	echo "....Done." 1>&2
 else
 	echo "Invalid combination of options." 1>&2
 	exit 1
